Feature/refactor string format (#17)
* Replaces all str.format occurrencies by f-string

f-string is a new way of formatting strings since Python 3.6
It's less verbose, so code was updated to use this new syntax.

* Updates README

Adds --max-line-length=127 flag to flake8 example.
The default value is 80, but 127 it's more reasonable for these
days. Also, 127 is the number of columns in a GitHub editor.

* Removes code smell

"Multiple commands in one line"

* Removes unnecessary try..except

platiagro/sdk always saves columns and featuretypes on metadata

* Adds unit tests that triggers character encoding detection

UTF-8 with 'BOM' files are instantly detected.
Other encodings, like Windows-1252 are not detected with 100%
accuracy, but the most likely is assumed as true.

* Fix code smells

- module Imported but unused

* Adds type hints to methods

* Limits the number of rows to peek during encoding detection

This process may take too much time on large datasets. So it's
healthy to set a limit.
Also, changes nrows default to 100.

* Renames function 'read_csv' to 'read_into_dataframe'

Makes clear it's not pandas.read_csv, but our custom function.