Adds sample dataset 'hotel_bookings'
This data set contains booking information for a city hotel and a
resort hotel, and includes information such as when the booking was made,
length of stay, the number of adults, children, and/or babies,
and the number of available parking spaces, among other things.
All personally identifying information has been removed from the data.

Downloaded from:
https://www.kaggle.com/jessemostipak/hotel-booking-demand